refactor: move other user agent logic into deviceUtils
