--- apps/journeys-admin/src/libs/deviceUtils/index.ts
@@ -1,1 +1,0 @@
-export { hasTouchScreen } from './deviceUtils'
