+++ libs/journeys/ui/src/components/VideoTrigger/VideoTrigger.tsx
@@ -3,6 +3,8 @@ import { useRouter } from 'next/router'
 import { ReactElement, useEffect, useState } from 'react'
 import Player from 'video.js/dist/types/player'
 
+import { isIPhone } from '@core/shared/ui/deviceUtils'
+
 import { handleAction } from '../../libs/action'
 import type { TreeBlock } from '../../libs/block'
 import { useJourney } from '../../libs/JourneyProvider'
@@ -14,17 +16,6 @@ type VideoTriggerProps = (
   | Pick<TreeBlock<VideoTriggerFields>, 'triggerAction' | 'triggerStart'>
 ) & {
   player?: Player
-}
-
-function iPhone(): boolean {
-  if (
-    typeof navigator === 'undefined' ||
-    typeof navigator.userAgent === 'undefined'
-  )
-    return false
-
-  const userAgent = navigator.userAgent
-  return userAgent.includes('iPhone')
 }
 
 export function VideoTrigger({
@@ -46,7 +37,7 @@ export function VideoTrigger({
           setTriggered(true)
           player.pause()
 
-          if (variant === 'embed' && !iPhone()) {
+          if (variant === 'embed' && !isIPhone()) {
             handleAction(router, triggerAction)
             return
           }
