+++ libs/shared/ui/src/libs/deviceUtils/index.ts
@@ -1,0 +1,1 @@
+export { hasTouchScreen, isIPhone, isMobile, isIOS } from './deviceUtils'
